Module DafapLayout
This module allows to manage layouts in a project.
We can define a particular layout for errors (404 and exceptions), an individual layout for each module or each controller, a default layout for all modules or conrollers for which we do not set specific layout.
Layouts accept parameters from a configuration file, specifying eg strings, css files, script files, images ...